backend/requirements.txt
Flask
psycopg2-binary
Flask-SQLAlchemy
flask-cors